refactor(application-templates-driving-license): splitting the draft state into separate files
