+++ libs/application/templates/driving-license/src/forms/draft/index.ts
@@ -1,0 +1,1 @@
+export { draft } from './draft'
